pyproject.toml
[project]
name = "ai-trader"
version = "0.2.0"
description = "AI-powered trading agent system with memory, sentiment analysis, and technical indicators"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangChain Core & Integrations (Latest Jan 2026)
    "langchain>=1.2.0",
    "langchain-core>=1.2.6",
    "langchain-openai>=1.1.6",
    "langchain-google-genai>=4.1.3",
    "langchain-mcp-adapters>=0.2.1",
    # MCP Server Framework
    "fastmcp>=2.14.2",
    # Market Data Sources
    "tushare>=1.4.24",
    "efinance>=0.5.5",
    # Core Dependencies
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.5",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "requests>=2.32.3",
    # Memory & Vector Search
    "sentence-transformers>=3.3.1",
    "chromadb>=0.5.23",
    # AI/ML Models
    "transformers>=4.48.0",
    "torch>=2.5.1",
    # Caching & Performance
    "cachetools>=5.5.0",
    # YAML Support
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = []
